backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
httpretty==1.1.4
idna==3.10
lxml==5.3.0
multidict==6.1.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-Wappalyzer==0.3.1
requests==2.32.3
rsa==4.9
setuptools==75.0.0
soupsieve==2.6
sqlparse==0.5.1
Sublist3r==1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wappalyzer==1.0
yarl==1.11.1
